+++ src/main/scala/com/eigenroute/portfolioanalysis/rebalancing/PortfolioWeightDifference.scala
@@ -1,3 +1,3 @@
 package com.eigenroute.portfolioanalysis.rebalancing
 
-case class PortfolioWeightDifference(eTFCode: ETFCode, weightDifference: Double)
+case class PortfolioWeightDifference(eTFCode: ETFCode, weightDifference: BigDecimal)
